add kafka events

+++ planner-core-api/src/main/kotlin/com/kvsinyuk/plannercoreapi/model/response/CreateProductResponseDto.kt
@@ -1,7 +1,9 @@
 package com.planner.dto.response
 
+import java.math.BigDecimal
+
 data class CreateProductResponseDto(
     val id: Long,
-    val price: Float,
+    val price: BigDecimal,
     val description: String? = null
 )
