+++ planner-core-api/src/main/kotlin/com/kvsinyuk/plannercoreapi/model/response/GetProductsResponseDto.kt
@@ -1,11 +1,13 @@
 package com.planner.dto.response
 
+import java.math.BigDecimal
+
 data class GetProductsResponseDto(
     val products: List<ProductResponseDto>
 )
 
 data class ProductResponseDto(
     val id: Long,
-    val price: Float,
+    val price: BigDecimal,
     val description: String
 )
